refactor: statistics 도메인 생성
- daily statistics를 statistics 도메인으로 이동
